+++ CHANGELOG.rst
@@ -5,8 +5,8 @@ Changelog
 0.1.16 (2026-06-02)
 -------------------
 
-* Add move progress API methods: ``get_all_move_progress``, ``get_move_progress``,
-  ``create_move_progress``, ``create_move_progress_batch``, ``update_move_progress``.
+* Add move status API methods: ``get_all_move_status``, ``get_move_status``,
+  ``start_move_batch``, ``update_move_status``.
 
 0.0.0 (2025-01-07)
 ------------------